# backend/docs/apidoc.yaml
openapi: 3.0.3
info:
  title: Beehive Monitoring API
  description: |
    API for ingesting beehive telemetry data and presenting it to authenticated users.
    This specification covers both internal (frontend-facing) and external (hardware-facing) APIs.
  version: 1.0.0

servers:
  - url: http://localhost:4000
    description: Local development

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessionId

    ingestKey:
      type: apiKey
      in: header
      name: X-INGEST-KEY

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string

    Hive:
      type: object
      properties:
        id:
          type: integer
        beekeeper_id:
          type: integer
        name:
          type: string
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Device:
      type: object
      properties:
        id:
          type: integer
        hive_id:
          type: integer
        installed_at:
          type: string
          format: date-time
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        active:
          type: boolean

    Reading:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: integer
        recorded_at:
          type: string
          format: date-time
        temperature_c:
          type: number
        battery_voltage:
          type: number
          nullable: true
        signal_strength:
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time

security:
  - sessionCookie: []

paths:
  /health:
    get:
      summary: Health check
      description: Confirms the server is running.
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time

  /auth/login:
    post:
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/logout:
    post:
      summary: Log out
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      summary: Get current user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"

  /sessions/current:
    get:
      summary: Get current session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    type: object
                    properties:
                      id:
                        type: integer
                      expiresAt:
                        type: string
                        format: date-time
                  user:
                    $ref: "#/components/schemas/User"

    delete:
      summary: Invalidate current session
      responses:
        "200":
          description: Session invalidated

  /hives:
    get:
      summary: List hives
      responses:
        "200":
          description: List of hives
          content:
            application/json:
              schema:
                type: object
                properties:
                  hives:
                    type: array
                    items:
                      $ref: "#/components/schemas/Hive"

    post:
      summary: Create hive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [name]
              properties:
                name:
                  type: string
                notes:
                  type: string
      responses:
        "201":
          description: Hive created
          content:
            application/json:
              schema:
                type: object
                properties:
                  hive:
                    $ref: "#/components/schemas/Hive"
        "400":
          description: Invalid input

  /hives/{hiveId}:
    get:
      summary: Get hive
      parameters:
        - name: hiveId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hive
          content:
            application/json:
              schema:
                type: object
                properties:
                  hive:
                    $ref: "#/components/schemas/Hive"
        "404":
          description: Hive not found

    put:
      summary: Update hive
      parameters:
        - name: hiveId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                notes:
                  type: string
      responses:
        "200":
          description: Hive updated
        "404":
          description: Hive not found

    delete:
      summary: Delete hive
      parameters:
        - name: hiveId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hive deleted

  /devices:
    get:
      summary: List devices
      responses:
        "200":
          description: Devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"

    post:
      summary: Create device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [hiveId]
              properties:
                hiveId:
                  type: integer
      responses:
        "201":
          description: Device created
        "404":
          description: Hive not found

  /devices/{deviceId}:
    get:
      summary: Get device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
        "404":
          description: Device not found

    put:
      summary: Update device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required: [active]
              properties:
                active:
                  type: boolean
      responses:
        "200":
          description: Device updated
        "404":
          description: Device not found

    delete:
      summary: Delete device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device deleted

  /readings:
    get:
      summary: Query readings
      parameters:
        - name: deviceId
          in: query
          schema:
            type: integer
        - name: hiveId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Readings
          content:
            application/json:
              schema:
                type: object
                properties:
                  readings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Reading"

  /readings/latest:
    get:
      summary: Get latest reading
      parameters:
        - name: deviceId
          in: query
          schema:
            type: integer
        - name: hiveId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Latest reading
        "404":
          description: No readings found

  /ingest/reading:
    post:
      summary: Ingest telemetry (external)
      security:
        - ingestKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [deviceId, recordedAt, temperatureC]
              properties:
                deviceId:
                  type: integer
                recordedAt:
                  type: string
                  format: date-time
                temperatureC:
                  type: number
                batteryVoltage:
                  type: number
                signalStrength:
                  type: number
      responses:
        "201":
          description: Reading ingested
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized
        "404":
          description: Device not found or inactive